refactor(platform): extract SceneEmbedViewer
The dashboard scene page and the fullscreen preview route each carried their
own PreviewModel with the same viewer props, the same resolveBakedShadowSource
memo, and the same spinners. Consolidate them before the /embed and /preview
routes split.

- add components/scene-embed/{scene-embed-viewer,scene-embed-info-popover}
- move use-preview-scene.ts to components/scene-embed/use-scene-embed-scene.ts
- drop both local PreviewModel copies and the inlined InfoPopover

No behavior change. The viewer container already defaults to h-full w-full, so
dropping that class in preview-fullscreen is a no-op. Routes, auth, and URLs
are untouched.
